Java Programming
Create a LinkedList class, write a method insert that inserts integers at the end
of the list. Write method that searches through the linked list and returns
whether or not a specific number exists in the list, a method that remove a
node from the linked list. Don’t allow duplicate elements.